backend/requirements.txt
asgiref==3.5.2
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-ckeditor==6.3.2
django-compat==1.0.15
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-sql-sniffer==1.0.3
django4-background-tasks==1.2.7
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
mailjet-rest==1.3.4
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
Pillow==9.3.0
psycopg2-binary==2.9.5
PyExcelerate==0.10.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.6
qrcode==7.3.1
requests==2.28.1
requests-futures==1.0.0
s3transfer==0.6.0
sendgrid==6.9.7
sentry-sdk==1.10.1
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
typing_extensions==4.4.0
urllib3==1.26.12
XlsxWriter==3.0.3
